Add StepVerifierOptions.copy(), withVirtualTime uses default vts
This commit adds a method to make a copy of a StepVerifierOptions. It
is used by `StepVerifier#withVirtualTime(Supplier, StepVerifierOptions)`
to make a copy and augment the provided options with a default supplier
of the VirtualTimeScheduler if none has been defined in the original
options.

This approach is preferable to throwing an NPE when one simply wants
to use default virtual time but change other unrelated specific options.