Prepare 3.5 behavioral change of take by aliasing to takeEager
This commit introduces the `takeEager` operator and makes `take` an
alias for it. `takeEager` is the new canonical way of obtaining the
current take behavior (unbounded upstream request, cancels when enough
elements have been emitted).

Furthermore, it reflects in `take` javadoc that the behavior will change
in 3.5.0 so that `take` becomes the new official name for `limitRequest`
behavior (still cancel whenever enough elements have been emitted,
but limit the upstream request to what is strictly necessary).
See #2690 for planned changes.

Fixes #2339.